squares now changing color
