+++ party-on/src/index.css
@@ -18,7 +18,7 @@ ol, ul {
 }
 
 .square {
-  background: #fff;
+  /* background: #fff; */
   border: 1px solid #999;
   float: left;
   font-size: 24px;
@@ -36,9 +36,9 @@ ol, ul {
   outline: none;
 }
 
-.kbd-navigation .square:focus {
+/* .kbd-navigation .square:focus {
   background: #ddd;
-}
+} */
 
 .game {
   display: flex;
